Wish 업데이트 기능 구현, 리팩토링

+++ src/main/java/com/project/wish/service/FileUploader.java
@@ -5,7 +5,6 @@ import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -14,12 +13,21 @@ import java.util.Objects;
 @Component
 public class FileUploader {
 
-    public String getUploadFilePath(MultipartFile uploadFile) throws IOException {
+    public String getUploadFilePath(MultipartFile uploadFile) {
         String fileName = StringUtils.cleanPath(Objects.requireNonNull(uploadFile.getOriginalFilename()));
         Path path = Paths.get("src/main/resources/static/uploadImages/" + fileName);
-        Files.write(path, uploadFile.getBytes());
+        System.out.println("FileUploader.getUploadFilePath : " + fileName);
 
-        return path.toString().replace("src/main/resources/static", "");
+        if (fileName.equals("")) {
+            return "/uploadImages/default.png";
+        }
+        try {
+            if (!Files.exists(path)) {
+                Files.write(path, uploadFile.getBytes());
+            }
+            return path.toString().replace("src/main/resources/static", "");
+        } catch (Exception e) {
+            return "/uploadImages/default.png";
+        }
     }
-
 }
